+++ packages/react-ui/components/DropdownMenu/__stories__/DropdownMenu.creevey.stories.tsx
@@ -1,13 +1,14 @@
-import React from 'react';
+import React, { useState } from 'react';
 
 import { delay } from '../../../lib/utils';
 import { Button } from '../../Button';
 import { CreeveyTests, Meta } from '../../../typings/stories';
 import { DropdownMenu } from '../DropdownMenu';
-import { MenuHeader } from '../../../components/MenuHeader';
-import { MenuItem } from '../../../components/MenuItem';
+import { MenuHeader } from '../../MenuHeader';
+import { MenuItem } from '../../MenuItem';
 import { OkIcon } from '../../../internal/icons/16px';
 import { PopupMenuDataTids } from '../../../internal/PopupMenu';
+import { ReactUIFeatureFlagsContext } from '../../../lib/featureFlagsContext';
 
 export default {
   title: 'DropdownMenu/Functional tests',
@@ -72,5 +73,56 @@ WithItemsAndIconsWithoutTextAlignment.parameters = {
   creevey: {
     tests: textAlignmentTests,
     skip: { 'themes dont affect logic': { in: /^(?!\bchrome\b)/ } },
+  },
+};
+
+const navigateInNestedMenuItems: CreeveyTests = {
+  async navigate() {
+    await this.browser
+      .actions({
+        bridge: true,
+      })
+      .click(this.browser.findElement({ css: `[data-tid~="${PopupMenuDataTids.caption}"]` }))
+      .sendKeys(this.keys.DOWN)
+      .sendKeys(this.keys.DOWN)
+      .perform();
+    const arrowDown = await this.browser.takeScreenshot();
+
+    await this.browser
+      .actions({
+        bridge: true,
+      })
+      .sendKeys(this.keys.ENTER)
+      .perform();
+    await delay(1000);
+    const enter = await this.browser.takeScreenshot();
+
+    await this.expect({ arrowDown, enter }).to.matchImages();
+  },
+};
+
+export const WithNestedMenuItems = () => {
+  const [caption, setCaption] = useState('not selected');
+  const onClick = () => {
+    setCaption('selected');
+  };
+  return (
+    <ReactUIFeatureFlagsContext.Provider value={{ menuItemsAtAnyLevel: true }}>
+      <DropdownMenu menuWidth="300px" caption={<Button use="primary">{caption}</Button>}>
+        <>
+          <div>
+            <MenuItem>Раз</MenuItem>
+            <MenuItem onClick={onClick}>Два</MenuItem>
+          </div>
+          <MenuItem>Три</MenuItem>
+        </>
+      </DropdownMenu>
+    </ReactUIFeatureFlagsContext.Provider>
+  );
+};
+WithNestedMenuItems.storyName = 'With nested menu items';
+WithNestedMenuItems.parameters = {
+  creevey: {
+    tests: navigateInNestedMenuItems,
   },
 };