+++ packages/react-ui/components/Select/Select.tsx
@@ -1,6 +1,7 @@
 import React, { ReactNode, ReactPortal, AriaAttributes } from 'react';
 import invariant from 'invariant';
 import { globalObject } from '@skbkontur/global-object';
+import debounce from 'lodash.debounce';
 
 import {
   isKeyArrowDown,
@@ -35,6 +36,11 @@ import { isTheme2022 } from '../../lib/theming/ThemeHelpers';
 import { ThemeFactory } from '../../lib/theming/ThemeFactory';
 import { MenuHeaderProps } from '../MenuHeader';
 import { SizeProp } from '../../lib/types/props';
+import {
+  getFullReactUIFlagsContext,
+  ReactUIFeatureFlags,
+  ReactUIFeatureFlagsContext,
+} from '../../lib/featureFlagsContext';
 
 import { ArrowDownIcon } from './ArrowDownIcon';
 import { Item } from './Item';
@@ -226,6 +232,7 @@ export class Select<TValue = {}, TItem = {}> extends React.Component<SelectProps
   private buttonElement: FocusableReactElement | null = null;
   private getProps = createPropsGetter(Select.defaultProps);
   private setRootNode!: TSetRootNode;
+  private featureFlags!: ReactUIFeatureFlags;
 
   public componentDidUpdate(_prevProps: SelectProps<TValue, TItem>, prevState: SelectState<TValue>) {
     if (!prevState.opened && this.state.opened) {
@@ -238,17 +245,24 @@ export class Select<TValue = {}, TItem = {}> extends React.Component<SelectProps
 
   public render() {
     return (
-      <ThemeContext.Consumer>
-        {(theme) => {
-          this.theme = ThemeFactory.create(
-            {
-              menuOffsetY: theme.selectMenuOffsetY,
-            },
-            theme,
+      <ReactUIFeatureFlagsContext.Consumer>
+        {(flags) => {
+          this.featureFlags = getFullReactUIFlagsContext(flags);
+          return (
+            <ThemeContext.Consumer>
+              {(theme) => {
+                this.theme = ThemeFactory.create(
+                  {
+                    menuOffsetY: theme.selectMenuOffsetY,
+                  },
+                  theme,
+                );
+                return <ThemeContext.Provider value={this.theme}>{this.renderMain()}</ThemeContext.Provider>;
+              }}
+            </ThemeContext.Consumer>
           );
-          return <ThemeContext.Provider value={this.theme}>{this.renderMain()}</ThemeContext.Provider>;
         }}
-      </ThemeContext.Consumer>
+      </ReactUIFeatureFlagsContext.Consumer>
     );
   }
 
@@ -488,7 +502,7 @@ export class Select<TValue = {}, TItem = {}> extends React.Component<SelectProps
   private getSearch = () => {
     return (
       <div className={styles.search()} onKeyDown={this.handleKey}>
-        <Input ref={this.focusInput} onValueChange={this.handleSearch} width="100%" />
+        <Input ref={this.debouncedFocusInput} onValueChange={this.handleSearch} width="100%" />
       </div>
     );
   };
@@ -516,7 +530,7 @@ export class Select<TValue = {}, TItem = {}> extends React.Component<SelectProps
       <Input
         autoFocus
         value={this.state.searchPattern}
-        ref={this.focusInput}
+        ref={this.debouncedFocusInput}
         onValueChange={this.handleSearch}
         width="100%"
       />
@@ -569,10 +583,10 @@ export class Select<TValue = {}, TItem = {}> extends React.Component<SelectProps
     return getRootNode(this);
   };
 
-  private focusInput = (input: Input) => {
-    // fix cases when an Input is rendered in portal
-    globalObject.setTimeout(() => input?.focus(), 0);
-  };
+  // fix cases when an Input is rendered in portal
+  // https://github.com/skbkontur/retail-ui/issues/1995
+  private focusInput = (input: Input) => input?.focus();
+  private debouncedFocusInput = debounce(this.focusInput);
 
   private refMenu = (menu: Menu) => {
     this.menu = menu;
@@ -624,8 +638,10 @@ export class Select<TValue = {}, TItem = {}> extends React.Component<SelectProps
   };
 
   private handleSearch = (value: string) => {
+    const menuItemsAtAnyLevel = this.featureFlags.menuItemsAtAnyLevel;
+
     this.setState({ searchPattern: value });
-    this.menu?.highlightItem(1);
+    this.menu?.highlightItem(menuItemsAtAnyLevel ? 0 : 1);
   };
 
   private select(value: TValue) {