+++ packages/react-ui/lib/featureFlagsContext/ReactUIFeatureFlagsContext.tsx
@@ -1,0 +1,20 @@
+import React from 'react';
+
+export interface ReactUIFeatureFlags {
+  tokenInputRemoveWhitespaceFromDefaultDelimiters?: boolean;
+  sidePageEnableFocusLockWhenBackgroundBlocked?: boolean;
+  spinnerLoaderRemoveDefaultCaption?: boolean;
+  menuItemsAtAnyLevel?: boolean;
+  textareaUseSafari17Workaround?: boolean;
+}
+
+export const reactUIFeatureFlagsDefault: ReactUIFeatureFlags = {
+  tokenInputRemoveWhitespaceFromDefaultDelimiters: false,
+  sidePageEnableFocusLockWhenBackgroundBlocked: false,
+  spinnerLoaderRemoveDefaultCaption: false,
+  menuItemsAtAnyLevel: false,
+  textareaUseSafari17Workaround: false,
+};
+
+export const ReactUIFeatureFlagsContext = React.createContext<ReactUIFeatureFlags>(reactUIFeatureFlagsDefault);
+ReactUIFeatureFlagsContext.displayName = 'ReactUIFeatureFlagsContext';