+++ packages/react-ui/lib/rootNode/getRootNode.ts
@@ -2,8 +2,9 @@
 import { findDOMNode } from 'react-dom';
 import React from 'react';
 import warning from 'warning';
-import { globalObject, isInstanceOf } from '@skbkontur/global-object';
+import { globalObject } from '@skbkontur/global-object';
 
+import { isInstanceOf } from '../../lib/isInstanceOf';
 import { Nullable } from '../../typings/utility-types';
 
 import { isInstanceWithRootNode } from './rootNodeDecorator';