+++ packages/react-ui-validations/CHANGELOG.md
@@ -3,6 +3,39 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.14.1](https://github.com/skbkontur/retail-ui/tree/master/packages/react-ui-validations/compare/react-ui-validations@1.14.0...react-ui-validations@1.14.1) (2024-01-25)
+
+
+### Bug Fixes
+
+* **validations:** export ValidationsFeatureFlagsContext ([#3345](https://github.com/skbkontur/retail-ui/tree/master/packages/react-ui-validations/issues/3345)) ([0c4fe51](https://github.com/skbkontur/retail-ui/tree/master/packages/react-ui-validations/commit/0c4fe51e4748e2e6588ab1ef8a772d792659b396))
+
+
+
+
+
+# [1.14.0](https://github.com/skbkontur/retail-ui/tree/master/packages/react-ui-validations/compare/react-ui-validations@1.13.1...react-ui-validations@1.14.0) (2023-12-15)
+
+
+### Features
+
+* **react-ui,validations:** add FeatureFlags contexts ([#3312](https://github.com/skbkontur/retail-ui/tree/master/packages/react-ui-validations/issues/3312)) ([508c803](https://github.com/skbkontur/retail-ui/tree/master/packages/react-ui-validations/commit/508c803018360a396bb693c9272e1dc2490e9a7b))
+
+
+
+
+
+## [1.13.1](https://github.com/skbkontur/retail-ui/tree/master/packages/react-ui-validations/compare/react-ui-validations@1.13.0...react-ui-validations@1.13.1) (2023-12-04)
+
+
+### Bug Fixes
+
+* **validations:** add $ as allowed symbol ([#3321](https://github.com/skbkontur/retail-ui/tree/master/packages/react-ui-validations/issues/3321)) ([c6a864e](https://github.com/skbkontur/retail-ui/tree/master/packages/react-ui-validations/commit/c6a864eb6cdfb90052f9d2e075f5a33b2cfb871b))
+
+
+
+
+
 # [1.13.0](https://github.com/skbkontur/retail-ui/tree/master/packages/react-ui-validations/compare/react-ui-validations@1.12.0...react-ui-validations@1.13.0) (2023-08-10)